backend/requirements.txt
braintree==4.11.0
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
psycopg2==2.9.1
